implement `indexer run` command
